lib/oelite: move injection of imports from base.bbclass to baker.py
Before, sometimes when a recipe was parsed, injections were done.
There are no need to do it more than once.

Also; since the code was placed in an anonymous function, it was not
completely clear when the code was infact run. A problem has appeared
on some machines if the name of the oe-lite folder was sdk32, Paint,
or MSSQL (There could be more..) and was placed in the homedir.
The problem is removed when this patch.

Signed-off-by: Esben Haabendal <eha@doredevelopment.dk>