class/auto-package-utils: Set AUTO_PACAGE_UTILS_RPROVIDES var
The AUTO_PACKAGE_UTILS_ALL can fx. be used to add all auto packaged utils
in a recipe to a specific package provided by that recipe (fx. ${PN})

In i2c-tools, binutils, and mtd-utils the PN package is empty, and
therefore, it may be utilized to RDEPEND on all; something that might
happen quite often :)