class/rstage: Add lower bound to when rstage is run.
It might happen that one wants to insert a file into the set of RDEPENDS
from SRC_URI. If so, this currently fails due to uncertainty of when rstage
is run. rstage is currently run before do_compile, meaning it might start
with rstage which then cant copy a file from SRC_URI to rstage dir,
since do_fetch hasnt been run.