crosstool-ng: Fixup variable handling to fix toolchain build
With BitBake, exported variables set to empty string is not exported.
In OE-lite v3, these are exported with an empty string value.

This commit therefore changes the assignment of empty strings to
a number of variables, to setting the unexport flag instead, as the
objective is to make the variables not exported, to avoid breaking
build of various toolchain components (at least glibc).