lib/oelite: Add support for emit and omit variable flags
Set emit flag to a list of tasks to emit var for. Set it to empty string
to not have it emitted for any tasks.

Set omit flag to a list of tasks not to emit var for.

If task is listed both in emit and omit flag, it will not be emitted.