lib/oelite: Change meaning of "unexport" flag
Instead of adding "unset VARNAME" to the shell function run script, the
VARNAME var is simply not added to the run script.

This way, unexport does _not_ unset variables that is specifically
imported from the build environment.  To get that effect, add the
specific unset commands to the shell function instead.