class/library: Improved support for recipes not starting with "lib"
Improved support for recipes not starting with "lib", and thus not having
PN the same as the library name.  These recipes will now get the proper
library name added to PROVIDES (both for PN and PN_dev).
If LIBRARY_NAME is set, it is taken as the full library name. Otherwise,
library name is taken as lib + PN.