class/package: Add support for PACKAGE_META_VARS
With PACKAGE_META_VARS, we now introduce a generic approach for putting
package specific metadata information into the package files (tar balls).

This will be used for providing FILE_PRIORITY support, and could in the
future be used for easing the task of implementing a bill-of-material
task, and perhaps even on-line package management support, by making
it possible to generate fx. RPM packages from the tar balls.

The binconfig class should be refactored to use this if possible.