class/stage: Add FILE_PRIORITY support
With FILE_PRIORITY, recipes can mark specific files with a priority level.
When a file conflict occurs while staging, a file with a higher priority
level is allowed to override files with lower priority.

A recipe can set FILE_PRIORITY variable directly, or a package specific
FILE_PRIORITY_{PN}-whatever variable.  The syntax is a space seperated
list of file priorities. A file priority is either a single integer,
setting the default priority of all files in the package, or a filename
(full path) followed by a colon and an integer specificying the priority
for that specific file.