lib/oelite: Add support for ${INCOMPATIBLE_RECIPES}
When building the cookbook, ${INCOMPATIBLE_RECIPES} (if defined) is
interpreted as a space separated list of recipe match expressions.

A recipe match expression can be a regular expression (must not contain
'_'). Recipes with ${PN} that matches one of these regular expressions
are not added to the cookbook.

If only specific ${PV} is to be matched, the other recipe match expression
format can be used, where first part is a regular expression, followed
by '_' and then a value matched against ${PV}.

A typical use for this feature is to set ${INCOMPATIBLE_RECIPES} in
a machine configuration file, skipping known incompatible recipes, such
as specific kernel and bootloader recipes, without having to modify the
meta layer containing these recipes.

(closes #278)