gcc: Temporary workaround for image_qa failures with libelf version check
This changes all gcc image_qa version checks from errors into warnings,
making it possible to build toolchains although we currently have an
issue with sometimes assuming libelf is needed although it actually is not.
In these cases ct-ng removes CT_LIBELF_VERSION when building.  The
version check therefore fails since it cant read a variable which isnt
there.

When the libelf issue is resolved, this commit should be reverted.