gcc: Make sure $MACHINE is not exported
The glibc configure script barfs out with something like
    .../ld: cannot represent machine `arm:pandaboard'

When MACHINE variable is exported.

We probably should add code to lib/oelite/function.py to handle
the case where whitelisted shell variables are exported, which then
ends up being exported to the shell task although it does not have
the export flag set.  We should add code to go through all whitelisted
variables, and clearing the export flag (as with export -n in bash)
so that we always get the export status as defined by the metadata.