lib/oelite: Add COMPATIBLE_IF_FLAGS feature
If recipe has COMPATIBLE_IF_FLAGS variable set, it will be checked
after post_recipe_parse, and any USE flags listed in it (space seperated)
will have to be set. If one or more of them is not set, the recipe
will be skipped (ie. not added to cookbook).

(closes #325)