class/package: Change PACKAGE_META_VARS semantics
Change the semantics so that when finding a PACKAGE_META_VARS value for
a package, the defaults is inherited from "parent packages".  With
"parent packages", I mean that fx. bar-y is a parent package to
bar-y-dev, as defined by the naming.  But bar-yxi is not a child of
bar-y.

This should be useful to reduce the amount of LICENSE_* variables that
needs to be defined for mixed license recipes.

(cherry picked from commit d36a0ac7bfa339178d6e47db6f1379369f9287e5)