fetch/url: Improve error handling of bad ingredients
When we have a bad local file do not raise on signature mishmatch, but
refetch..

Fx a Ctrl-c when fetching results in a bad localfile and manual
intervention needed. With this fix a rebuild should be enough. If the
remote url is indeed bad the Fetch function still returns False.