class/package: Improved PACKAGE_META_VARS handling
Write not only the value, but also the relative change "level" for each
package.

The idea is that packages is organized in a tree, fx.
  foobar
    foobar-foo
      foobar-foo-dev
      foobar-foo-dbg
    foobar-bar
      foobar-foo-x
        foobar-foo-x-y
        foobar-foo-x-z

And the level of a PACKAGE_META_VARS value for a package is equal to
the level of its parent package, and incremented by one a package
specific value is set for the package.

The meta files generated will have two lines, with the value in first
line and level in the second line.

(cherry picked from commit 8a401a22200a949d9bcfc07f0bab6bf90e99c470)