lib/oelite: SVN fetcher
This implementation lacks a proper method for SVN authorization.

The current implementation re-uses stored authorization, which is the
expected normal usage.

When fetching from an SVN repository that requires authorization, which
is not stored, the SVN fetcher prompts for username and password, and
stores together with the normal if it works out.

The problem is that this only works properly together with "-d" option.
Without -d, the prompt for username is not visible, although you can
(and must) still enter username, and then password.
lib/oelite: Add scmdata=keep parameter support in SVN fetcher