lib/oelite: Add URL_WHITELIST / URL_BLACKLIST feature
By setting URL_WHITELIST and/or URL_BLACKLIST, you will be able to
control which remote url's that the OE-lite fetchers will be allowed
to fetch.

If URL_BLACKLIST matches, the url is not allowed.
If URL_WHITELIST is defined and does not match, the url is not allowed.
Otherwise, the url is allowed.

The syntax for both these variables is a space separated list of
regular expressions, which will be matched on the beginning of the urls.

They will be matched on the real url's, and not the OE-lite SRC_URI
values. For SRC_URI values using a protocol parameter (fx. git and svn),
this means that for
  git://oe-lite.org/foobar.git;protocol=http
The value that will be matched against URL_BLACKLIST/URL_WHITELIST is
  http://oe-lite.org/foobar.git

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>

(cherry picked from commit 47450e8f3546970207ce7b5413615293aab6b6e8)
(closes #256)