lib/oelite: Add PREMIRROR/MIRROR support for svn sources
Add support for PREMIRROR and MIRROR fetching of svn snapshot tarballs
for non-branch and non-scmdata=keep SVN sources.

Entries (lines) in PREMIRROR and MIRROR should be in the form
  svn://.*/  http://oe-lite.org/mirror/${INGREDIENTS_SUBDIR}/svn//

Notice the two trailing slashes ('//').  It indicates that a snapshot
tarball generated by the OE-lite mirror task will be fetched from
the specified mirror url/directory.

Any protocol supported by the url (urlgrab) fetcher is allowed,
ie. http/ftp/file and so on.

(refs #306)

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>