lib/oelite: Add PREMIRROR/MIRROR support for git sources
With this, PREMIRROR and MIRROR can be used for setting pre and post
mirrors for git sources.

If the mirror specification ends with "//" (double slash), the mirror
url will be completed with a git repository name as generated by the
mirror task (oe bake -t mirror). Fx.
  git://.*/  http://oe-lite.org/mirror/${INGREDIENTS_SUBDIR}/git//
which will try to use something like
http://oe-lite.org/mirror/foobar/git/git_oe-lite.org_foobar.git
as a mirror for git://oe-lite.org/foobar.git

Otherwise, the mirror url will be completed with the non-matched part
of the original url. Fx.
  git://oe-lite.org/ http://oe-lite.org/foobar/
which will try to use http://oe-lite.org/foobar/foobar.git as a mirror
for git://oe-lite.org/foobar.git

(refs #306)

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>