crosstool-ng: Refactor libc/version USE flags
Simplify the USE flags used to control selection of libc and libc versions.

Machine libc configuration is managed by
    USE_machine_libc
which should be set to values like
    USE_machine_libc = "glibc:2.14"
    USE_machine_libc = "eglibc:2_16:HEAD"
    USE_machine_libc = "uclibc:0.9.33.2"
    USE_machine_libc = "mingw:3.18"
And thus selects both which libc is to be used and the version.
The USE_machine_libc flag is expected to be set in machine configuration.

Configuration of libc for SDK hosts are managed by the following flags
    USE_sdk_linux_libc
    USE_sdk_mingw_libc
These should be set to values like
    USE_sdk_linux_libc = "glibc:2.12.2"
    USE_sdk_linux_libc = "eglibc:2_16:HEAD"
    USE_sdk_mingw_libc = "mingw:3.18"

If these USE flags are not set, the old USE flags works as before
    USE_toolchain_libc
    USE_toolchain_libc_version
    USE_toolcahin_libc_eglibc_revision
    USE_toolchain_mingwrt_version

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>

(cherry picked from commit 818bebe33c6df58a87bb85674b092a66aacecb31)