conf/overrides: Fix wildcard overrides generation
When fx. setting SDK_OS=linux, we should not be checking the SDK_OS
directly, but rather the HOST_OS/TARGET_OS variables instead, as they
will be config.sub'ed to fx. linux-gnu, and the correct wildcard overrides
will be set.

Of-course, for this to work, the late_overrides_fixup needs to run after
arch_update which sets the HOST_OS/TARGET_OS.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>