lib/oelite: New manifest command: cherry
Manifest command for assisting maintainers handling cherry picking
to release branches.

Example: work is committed to master branch.  Maintainer needs to finish
up version 3.2.1 for release. This is done on the 3.2 branch. By using
the "oe cherry master 3.2" command, the maintainer can get a list of
which commits are possible candidates for this release.

The list includes all commits which are not already included, and which
are not marked for a later release branch (in the example above, fx. 4
or 3.3).  Target version marks are saved as git notes.

An interactive mode is also implemented, which allows setting target
version for all unmarked commit candidates.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>