class/kernel: Improve default handling of entry point and load address
Instead of having a hard-coded default value (20008000) for both
USE_kernel_uimage_entrypoint and USE_kernel_uimage_loadaddress, default to
not setting these USE flags and query the value from the vmlinux file
instead (Use the _start symbol for both).

With this change, you only have to set the USE_kernel_uimage_entrypoint in
case the _start symbol for some reason is not giving the desired result.
The USE_kernel_uimage_loadaddress should only be set when it must differ
from the entry point address.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>