lib/oe-lite: build_ and sdk_ prefix to vendor part of arch tuple
Add build_ and sdk_ prefix to the vendor part of arch tuples.
This is needed to avoid configure scripts messing up when two (or three)
of build, sdk and machine is identical.  This should both prevent
configure scripts in wrongly not detect cross-compilation mode, and
for using the wrong compiler tools because of identical arch prefix.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>