class/arch: Remove MACHINE variable for common machine recipes
Machine, cross and canadian-cross recipes that is not MACHINE specific, will
get the MACHINE variable removed.  This avoids rebuilds of fx. cross:gcc when
switching between machines with same MACHINE_CPU.  And, it would of-course be
a bug if any recipe that was build as not being MACHINE specific would
actually use the MACHINE variable for anything.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>