class/package-qa: Do various QA checks on packages produced.
For now, it is mostly ELF related checks that are implemented, but this
class should be extended with other types of packaging QA checks.

Currently, the following checks are performed:

* Are ELF files in the correct dirs?  Libraries in designated libdirs, and
  other ELF files in designated bindirs.
* Does all library packages have proper lib* provides in PROVIDES_*?
* Does all library packages come with an ELF .so file with the correct
  SONAME?
* Is needed library dependencies in DEPENDS_*?
* Does DEPENDS_* have unneeded library dependencies?
* Do machine packages have ELF files with undesired DT_RPATH/DT_RUNPATH
  tags?
* Does ELF files come with absolute DT_RPATH/DT_RUNPATH tags instead of
  $ORIGIN relative?

The behaviour of these checks can be to some extend be controlled by
setting the 'qa' flag on the related variable.

To run package QA on fx. zlib, run

    oe bake zlib -t packageqa

To run package QA on all recipes involved in building base-rootfs, run

    oe bake base-rootfs -t packageqaall

See the reference manual for more details (TBD).

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>