lib/oelite: Ignore broken dependencies of produced packages
This commit allows building of recipes which produce packages that has
dependencies that cannot be satisfied.  Of-course, the packages cannot be
used (staged), but the recipe as such can be built. This was also the
behavior before package-qa was implemented.

It might make sense to add a check for such broken dependencies in
package-qa.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>