lib/oelite: Add oe setup command
The setup command can show/execute the command needed to prepare a
host for OE-lite. A file for each supported host configuration must
be added to the setup directory in the core metadata layer. The file
should contain the command for installing the dependencies needed
for using oe-lite. The file name should be:
'distro_release' as defined by the lsb_release tool. If no release
exist it should be "none"
The lsb_release tool is used for automatic host detection. A
list of the known host configurations can be shown using 'oe setup -l'
If only the setup command for the current host should be printed,
run 'oe setup -p'

Using this tool we can let the needed dependencies for oe-lite
be recorded along meta/core versions.

It is also possible that we need to explitly say that some host
configuration will not work with oe-lite. To do this, add a host
configuration file for that host with e.g.: 'echo "This host is not
supported by this version of OE-lite. Use 'oe setup -l' to get a
list of known host configurations"'

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>

(cherry picked from commit e103462542061154c69cfb9ab6cfa20faecf64c8)