lib/oelite: Rewrite of git fetcher
Complete rewrite of git fetcher, using only subprocess and git shell
command.  This way we can get rid of the need for using python-fetching
and GitPython, and should hopefully make it easier to avoid breaking
on every odd git version with changed behaviour.

The 'repo' and 'remote' git SRC_URI parameters are no longer needed/supported,
but can be left in for backwards compatibility.

The cherry command still needs to be rewritten to get rid of GitPython.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>