class/useflags: Remove MACHINE_USE_ for sdk, native, sdk-cross recipes
MACHINE_USE_ flags must not be used in sdk, native and sdk-cross recipes so
delete them before setting the USE_ flags.

Fixes bug where EXTRA_ARCH was set to ".${MACHINE}" in useflags.oeclass when a
MACHINE_USE flag was set in a sdk-cross recipe. In this case of sdk-cross:gcc
and use of MACHINE_USE_toolchain_min_kernel_machine.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>