Revert "Parallel parse (and steps towards parallel build)"
This reverts commit 33e02dd0dc096cfbd370858c4b3149f35691c62f.

Parallel parse is currently broken, so it needs to be reverted until it is
fixed.

The problem observed is that recipe metadata (DEPENDS_${PN}, RDEPENDS_${PN}
and other package dependency variables are modified by runq. But with the
current parallel parse implementation, the runq works on recipe summary
information, and not the real metadata, so changes (addition of the
'__provides' flag) are lost when the real recipe metadata is loaded from
cache.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>