Issue #27: ply compatibility
PLY >= 3.6 uses a set instead of a dict for the lextokens. It's just used as a list in the end, so check the type of the return value to determine how to convert the data.

(cherry picked from commit 18b7b4bea9a9315b29c5aa7340a03bcd45c98b4b)