conf/provided: Align minimal.conf, default.conf and all.conf
By rewriting all.conf to amend default.conf, and default.conf to amend
minimal.conf, it should be easier to make consistent changes to these
in the future.

As a consequence of this change, native:zip is dropped from minimal.conf
and default.conf, as it was already removed from all.conf
(see 367b92d25983bf48647e4528897aa2d3ddde9425).

Signed-off-by: Esben Haabendal <esben@haabendal.dk>