gcc, glibc: fix bootstrapping -fstack-protector detection and use
The glibc recipe unconditionally disables use of
-fstack-protector. That's a bit unfortunate, because upstream glibc
tries to use that for some security-critical components
(e.g. resolv). It does not enable it for all of glibc, as that would
imply a significant performance loss.

The reason for this disabling is not obvious from the git history, but
it was presumably done as a hack to simply make glibc build (or
rather: link). However, the problem is really in gcc: gcc, at the
linking stage, needs to know whether the target libc has the
infrastructure to support the -fstack-protector instrumentation. But
the autodetect in the configure stage of gcc cannot possibly work when
the target glibc doesn't even exist yet. So we need to make sure that
the configure variable gcc_cv_libc_provides_ssp is set to yes at all
stages (including bootstrap).

Starting with gcc 4.9, gcc's configure supports
--with-glibc-version=M.N which can be used to explicitly set the
target glibc version
(https://gcc.gnu.org/ml/gcc-patches/2013-11/msg00619.html). Ideally,
we should set M.N automagically from the glibc version that we're
actually building, but for now we do something a little less fancy: a
USE flag which can then be overridden from the glibc recipe.

While at it, remove the disabling of -fstack-protector from glibc.