image: Package blacklist for rstage
This add support for blacklisting of packages for rstage, which has been
used since at least 2011 to handle differences between differences in
dependencies in rootfs and initramfs.

Ideally, we should have support for using multiple distros (or something
like that), so that initramfs and rootfs could use different distro
configuration, while still being build as a single image.

But until then, this seems like the best solution we have been able to
come up with, so let's get it merged.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>