oelite: DictMeta: reduce __flag_index memory consumption
A dict uses at least 280 bytes, while a 5-element list is about 80
bytes. So change the __flag_index member from a dict with a fixed set
of keys to a list, indexed by keys determined from the class member
INDEXED_FLAGS.

Also, a set itself is also 280 bytes, and since these are often empty,
we can use None instead.

The addition in get_flag is a temporary hack - arguably,
"__flag_index" shouldn't be in the list returned by .keys(). But I'll
get rid of this in another way in a later patch.

This saves a whopping 0.5 MiB out of 350 MiB. Far from impressive, but
it's a start...