oelite: DictMeta: force string interning when doing deep copies
The deep-copying method in the builtin copy module just returns the
string itself (with an elevated refcount, obviously). While that is
simple, fast and safe, we can do a little better by hooking into the
deepcopying and force interning of the string - that will then reduce
the overall memory consumption when equal-but-not-identical strings
are used as keys (or values) in various dicts.

Concretely, this cuts the memory consumption in my test case by 6.5
MiB, from 349.8 MiB to 343.3MiB (with the total number of string
objects dropping from 827498 to 692265). Looking at the string objects
in the dumped .json file reveals that there are probably other places
in the code base where slapping an intern() might help - for example,
there are 5390 copies of the string "stage_base_prefix" (each
consuming 54 bytes), 5238 copies of "BUILD_PREFIX", etc., etc.

While this obviously makes the deep-copying slightly slower, one also
gains some performance by simply not allocating that much memory, and
subsequent string comparisons, e.g. when doing hash table lookups, may
also become faster (when two string objects are both interned, python
doesn't need to actually compare the string contents to establish a
match).