oelite: lex/parse: slap some interns here and there
No, this is not about extremely poor employee management practices...

We can save quite a bit of memory by interning the strings used for
storing e.g. variable names. Doing this as early as possible (that is,
at the lexing stage) means that most subsequent copies of these
strings (e.g. when used as keys in a dict) will still refer to the
single interned string.

Unfortunately, there are still many cases where "long-lived" strings
are constructed subsequently as a concatenation or substrings of other
strings, but I'm in the process of trying to track those down.

For the test case I'm using while doing this
memory-footprint-reduction-exercise, the memory use reported by meliae
drops from 343.3MiB to 329.8MiB (delta 13.5MiB), with 692265 string
objects before and 488359 after. The cold-cache case (where we're not
using any pickled data) is similar, with memory use dropping from
293.1MiB to 278.7MiB and the number of string objects from 353368 to
107374.

Yes, these numbers also show that we're using much more memory when
we're reading pickled recipes. I'm going to attack that next.

It obviously costs a little extra time to do the lookup(/+insert), but
that is more than compensated for by the fewer trips into the kernel
for allocating more memory. Also, some string comparisons will get
cheaper (for two interned strings, equality is the same as
identity). This isn't just speculation; I consistently get slightly
shorter runtimes with this patch applied.