oelite: DictMeta: distinguish "simple" and "complex" variables, prep
In order to get rid of the many many dicts with only the empty string
as key, let's try to distinguish "simple" and "complex"
variables. We'll do this by making DictMeta instances have an extra
member, smpl, which will hold the values of the simple variables
directly. As preparation, rename the current dict member to cplx.

I choose this four-letter abbreviation for two reasons. First, it
makes a patch much simpler to read when one identifier is replaced by
another of the same length. Second, this patch gets rid of the
ugliness of using a builtin identifier (dict) as a member name, and
thus makes the syntax highlighting in my editor much more useful -
that goal wouldn't be achieved if we used 'complex'.