oelite: DictMeta: distinguish simple and complex variables
A simple variable is one with no flags or anything else attached. We
save 27% of our memory use by storing these simple variables in a dict
of their own. It does require some more bookkeeping code to ensure
that we promote a simple variable to a complex one whenever a flag is
attached, and to ensure that we look in both .smpl and .cplx when
relevant, but the memory saving is really worth it, IMO.

Total 907746 objects, 162 types, Total size = 265.4MiB (278284648 bytes)
 Index   Count   %      Size   % Cum     Max Kind
     0  477849  52 205125720  73  73  196888 dict
     1  109159  12  41175384  14  88    4328 set
     2  188808  20  13768624   4  93     456 tuple
     3   71836   7   7746300   2  96   44688 str

Total 684628 objects, 162 types, Total size = 192.5MiB (201898106 bytes)
 Index   Count   %      Size   % Cum     Max Kind
     0  253749  37 128578296  63  63  196888 dict
     1  109159  15  41175384  20  84    4328 set
     2  188808  27  13768624   6  90     456 tuple
     3   71775  10   7742322   3  94   44688 str