oelite: DictMeta: don't expose the __flag_index
The __flag_index is arguably an implementation detail of DictMeta, so
"__flag_index" shouldn't be in the list returned by .keys() - that
just causes various pain when we then have to deal with users
iterating over all keys and fetching the corresponding values.

So make the flag index a direct member of the DictMeta instance. This
is mostly done by s/dict["__flag_index"]/__flag_index/g.