oelite: cookbook.py: get str objects from sqlite, not unicode
When we send strings on a detour through the in-memory sqlite
database, we get them back as unicode strings. That's almost certainly
not what we want, and prevents us from using intern() in various
places to save memory. Even if we didn't try to intern strings,
unicode objects waste memory simply by using 4 bytes per
character. The documented way to get TEXT columns back out as str
objects is to set .text_factory, so let's use that.