oelite: baker.py: always print dependency tree and hash computation times
It's quite useful to get a breakdown of where oelite actually spends
its time. Using -d enables way too much, so these lines would vanish in
the sea of debug output. It might be useful to introduce a
'--debug-lite' controlling the stuff that just generates a single line
here and there, but for now just enable this unconditionally.