oelite: MetaData: make MetaData.builtin_nohash a frozenset
Since we're doing a lookup in MetaData.builtin_nohash for each and
every variable we're dumping (to check whether we shouldn't), it's a
good idea if the data structure allows lookup faster than O(n) -
especially since most variables are indeed _not_ in this list. Since
this list is a compile-time constant and nothing adds to the list, we
might as well use a frozenset.