oelite: MetaData: do less flag lookups during metadata hash computation
Early in dump_var, we get a copy of the flag => value mapping
associated to var. We might as well hang on to that a little longer
and use that dict for looking up the existence/value of certain flags
instead of calling .get_flag - dict lookups are fast, method calls not
so. This alone speeds up the hash computation by about 8%.