oelite: MetaDict: expose and use method for filling expand cache
This is a really ugly hack, but it cuts the metadata hash computation
time by a third. The idea is to ensure that we've actually done a
lookup of every variable in a recipe's metadata, thus populating the
expand cache, before copying the metadata to the recipe's tasks:

When we copy the DictMeta, we only create a shallow copy of the
expand_cache. This is perfectly fine, since any subsequent changes to
any variable in the DictMeta will just cause that variable (and its
dependents) to be deleted from the copied expand_cache - in effect,
what we rely on is that the values in the expand cache are immutable.

When we compute the task metadata hashes, we will end up iterating
over all keys and looking up their values. By ensuring that all values
have been looked up once and cached, we avoid doing the same expensive
expansions for each task.

We obviously only want to do this once per recipe, but the most
natural place to ensure that we get a populated expand_cache is just
before we copy the metadata in task.py. Thus I had to add a boolean
flag to indicate whether this DictMeta's _fill_expand_cache has
already been called - but this is admittedly rather ugly - calling
._fill_expand_cache too early and we lose the potential benefits.