oelite: runq.py: add some timing_info calls to prune_runq_*
An "oe bake world" usually ends up spending a lot of time in (at least
one of) these. Turning debug info on isn't helpful since the messages
would drown. So unconditionally print this piece of info.

If this is too chatty, one could maybe add a min_time parameter
(defaulting to 0) to timing_info, and ignore the call if the duration
was less than the min_time. These calls would then grow a
min_time=5 argument for some suitable value of 5.