lib/oelite/meta/meta.py: move some string building out of tight loop
The string.replace line is called millions of times. We might as well
build the replacement string while building the dynvars list (which is
only done a few thousand times) to save lots of string formatting
calls.

While at it, fix the "prototype" of dump_var: it obviously expects a
list of pairs, not a dict.