oelite: runq.py: add an index to the recdepend table
I think the use of an in-memory sql database makes the code rather
hard to understand, maintain and optimize (because of all the double
bookkeeping between python data structures and the database). I also
believe that it should at some point be ripped out.

However, since that's what we have right now, we might as well make
full use of it. Profiling shows that we spend a lot of time in
get_recdepends(). We're doing 100000s of SELECT queries against the
runq.recdepend table with ~10000 rows, so that's not really
surprising, since we probably end up doing a linear walk over the
table each time.

Creating an index cuts over 75% of the "Building dependency tree
time", from ~6m04s to ~1m29s.