classes/binconfig-stage: create destination dirs before symlinking
When staging packages that inherit binconfig, binconfig_stage_fixup() is
called to create a symlink:
./stage/cross/${bindir}/${PN}-config -> ./stage/machine/${bindir}/${PN}-config

When binconfig_stage_fixup() is called the destination directory for the
symlink (./stage/cross/${bindir}) might not exist, in which case the
symlink creating fails. This is the case when the first dependency being
stages uses binconfig-fixup.

Simply fix this by always creating the destination directory before
symlinking.