conf/provided: Make default usable again
As fakeroot as of 53e6e982456e depends on libacl1-dev, and we do not
have a recipe for this, default is not usable at all without it.

For conf/provided/all.conf, libacl1-dev was added in 367b92d25983b
(although the PR comments suggested that it was moved to another commit).

Signed-off-by: Esben Haabendal <esben@haabendal.dk>