lib/oelite: slightly sa[nf]er dynvar handling
The current logic blindly replaces any occurrence in any variable of
the value of ${TOPDIR} by the literal string '${TOPDIR}', even if the
variable just happened to contain that substring by coincidence (this
could happen if someone has the oe-lite manifest in something like
"/build" - a string which is not entirely unlikely to occur
elsewhere).

Since get() is a thin wrapper for _get(), it's not actually more
expensive to get the set of dependencies for a given variable. A saner
heuristic is to only apply the dynvar replacement to variables that
actually depend on the dynvar.

Speed-wise, the logic is slightly more complicated, but OTOH, for
simple variables with no dependencies we get to skip the six
string.replace calls, so I think it should even out.

Since a variable is not in its own dependency set, "oe show" now no
longer displays

# DATETIME='${DATE}${TIME}'
DATETIME='${DATETIME}'

but rather the slightly more useful

# DATETIME='${DATE}${TIME}'
DATETIME='20160706123705'

Since all the dynvars have the nohash flag (or are listed in
builtin_nohash), this in itself should not cause any problems -
rather, I think it's a nice side effect.

This doesn't mean that this patch doesn't change metadata hashes at
all. This is because there are some non-nohash variables which do not
explicitly depend on $TOPDIR/$WORKDIR but, by their construction, do
contain their values. Diffing the output of "oe show glibc --nohash"
before and after this patch shows that at least these variables are
affected:

FILESPATH
FILESPATHBASE
FILESPATH_EXISTS
FILE_DIRNAME
SIGNATURE_FILE

I'll treat these in a followup patch.