lib/oelite: add a few variables to builtin_nohash
After we only do dynvar replacement on variables explicitly depending
on the dynvar, there are a few variables whose contribution to the
task hash computation now depends on the absolute path of the
manifest. This makes prebakes almost useless.

The variables affected are

    FILESPATH
    FILESPATH_EXISTS
    FILE_DIRNAME
    SIGNATURE_FILE

However, we already have

conf/oelayout.conf:65:FILESPATH[emit] = ""

so FILESPATH doesn't contribute to the hash of any task.

For the remaining three, I see no reason for them to be part of the hash
computation:

FILE_DIRNAME and SIGNATURE_FILE are directly derived from the FILE
variable. So these two used to contribute values like
'${TOPDIR}/relative/path' and
'${TOPDIR}/relative/path/recipe.sig'. Differences in the 'recipe' part
are already reflected in different package names and/or versions. So
what these might have contributed is the 'relative/path' part. But if
the placement of a recipe relative to the manifest should contribute to
the hash, I'd argue that it is wrong for FILE itself to be nohash in the
first place. (Also, INGREDIENTS_SUBDIR is, usually, the basename of
FILE_DIRNAME and does contribute to the hash.)

FILESPATH_EXISTS is just an optimization to search fewer directories -
the local files we find still contribute their md5sum. So making this
nohash means that one might actually do a small recipe refactoring
without changing hashes, e.g. moving a patch file from /foo-1.2.3 to
/foo (removing the former and creating the latter directory in the
process).